Pre-setup and post-setup tasks
All executable files present in /entrypoint-tasks/pre-setup/ and
/entrypoint-tasks/post-setup/ are run at the beginning
or at the end, respectively, of the actual entrypoint-script.

Each of the executed files receive the CMD given to the container as their
arguments. This allows e.g. to change a plugin's config, or install a
technical requirement of a plugin that is to be installed.

This feature is not implemented in the images for Roundcube v1.5, since
they are not well maintained and will be dropped soon, anyways.