refactor check for non-empty workdir
introduce new envvar ROUNDCUBEMAIL_ALLOW_DIRTY_DOCROOT

if it is true, skip checks for empty docroot.
if it is false and the docroot is not empty, abort.
if it is empty and the docroot is not empty, print a warning and sleep,
giving the user the chance to manually abort.

this keeps backwards compatibility with existing deployments (that do not
have the envvar set)

the sleep timeout has been raise to 60seconds, so the user has more time to
manually abort (in a autostart context this will add an additional delay,
but shouldn't actually be harmful)

Closes: https://github.com/roundcube/roundcubemail-docker/issues/384